Move to_stream function to streams module that is newly created
